+++ daemon/app/GHCSpecter/Render/Web/Timing.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
 
 module GHCSpecter.Render.Web.Timing (
@@ -11,29 +12,44 @@ import Concur.Replica (
   onChange,
   onClick,
   onInput,
+  onMouseEnter,
+  onMouseLeave,
   style,
   width,
  )
 import Concur.Replica.DOM.Props qualified as DP (checked, name, type_)
-import Control.Lens (to, (^.), _1)
+import Concur.Replica.SVG.Props qualified as SP
+import Control.Lens (to, (^.), _1, _2)
 import Data.List qualified as L
-import Data.Maybe (fromMaybe, isNothing)
+import Data.Map.Strict qualified as M
+import Data.Maybe (catMaybes, fromMaybe, isNothing, maybeToList)
 import Data.Text qualified as T
 import Data.Time.Clock (
   secondsToNominalDiffTime,
  )
+import GHCSpecter.Channel.Common.Types (DriverId, ModuleName)
 import GHCSpecter.Channel.Outbound.Types (
   ModuleGraphInfo (..),
   SessionInfo (..),
  )
-import GHCSpecter.Data.Map (backwardLookup)
+import GHCSpecter.Data.Map (
+  BiKeyMap,
+  backwardLookup,
+ )
 import GHCSpecter.Data.Timing.Types (
   HasPipelineInfo (..),
   HasTimingTable (..),
+  TimingTable,
  )
-import GHCSpecter.Render.Web.ModuleGraph qualified as ModuleGraph
+import GHCSpecter.Graphics.DSL (
+  HitEvent (..),
+  Scene (..),
+  ViewPort (..),
+ )
 import GHCSpecter.Render.Components.TimingView qualified as TimingView
-import GHCSpecter.Render.Util (divClass)
+import GHCSpecter.Render.ConcurReplicaSVG (renderPrimitive)
+import GHCSpecter.Render.Util (divClass, xmlns)
+import GHCSpecter.Render.Web.ModuleGraph qualified as ModuleGraph
 import GHCSpecter.Server.Types (
   HasModuleGraphState (..),
   HasServerState (..),
@@ -44,29 +60,209 @@ import GHCSpecter.Server.Types (
 import GHCSpecter.UI.ConcurReplica.DOM (
   button,
   div,
+  hr,
   input,
   label,
+  p,
   text,
  )
+import GHCSpecter.UI.ConcurReplica.DOM.Events (
+  onMouseDown,
+  onMouseMove,
+  onMouseUp,
+ )
+import GHCSpecter.UI.ConcurReplica.SVG qualified as S
 import GHCSpecter.UI.ConcurReplica.Types (IHTML)
 import GHCSpecter.UI.Constants (
   timingHeight,
+  timingRangeHeight,
   timingWidth,
   widgetHeight,
  )
 import GHCSpecter.UI.Types (
   HasTimingUI (..),
   HasUIModel (..),
+  HasViewPortInfo (..),
   TimingUI,
   UIModel,
  )
 import GHCSpecter.UI.Types.Event (
+  BackgroundEvent (RefreshUI),
   BlockerDetailLevel (..),
   BlockerModuleGraphEvent (..),
   Event (..),
+  MouseEvent (..),
   TimingEvent (..),
  )
 import Prelude hiding (div, span)
+
+renderTimingChart ::
+  BiKeyMap DriverId ModuleName ->
+  TimingUI ->
+  TimingTable ->
+  Widget IHTML Event
+renderTimingChart drvModMap tui ttable =
+  S.svg
+    svgProps
+    [ S.style [] [text ".small { font: 5px sans-serif; } text { user-select: none; }"]
+    , S.g [] (fmap (renderPrimitive handler) rexp)
+    ]
+  where
+    handler hitEvent =
+      catMaybes
+        [ fmap (\ev -> TimingEv ev <$ onMouseEnter) (hitEventHoverOn hitEvent)
+        , fmap (\ev -> TimingEv ev <$ onMouseLeave) (hitEventHoverOn hitEvent)
+        ]
+    scene = TimingView.compileTimingChart drvModMap tui ttable
+    rexp = sceneElements scene
+    vpi = tui ^. timingUIViewPort
+    vp = fromMaybe (vpi ^. vpViewPort) (vpi ^. vpTempViewPort)
+    svgProps =
+      let viewboxProp =
+            SP.viewBox . T.intercalate " " . fmap (T.pack . show . floor @Double @Int) $
+              [ topLeft vp ^. _1
+              , topLeft vp ^. _2
+              , timingWidth
+              , timingHeight
+              ]
+          prop1 =
+            [ MouseEv . MouseDown <$> onMouseDown
+            , MouseEv . MouseUp <$> onMouseUp
+            , width (T.pack (show (timingWidth :: Int)))
+            , height (T.pack (show (timingHeight :: Int)))
+            , viewboxProp
+            , SP.version "1.1"
+            , xmlns
+            ]
+          mouseMove
+            | tui ^. timingUIHandleMouseMove =
+                [(\case Nothing -> BkgEv RefreshUI; Just xy -> MouseEv (MouseMove xy)) <$> onMouseMove]
+            | otherwise = []
+       in mouseMove ++ prop1
+
+renderMemChart ::
+  BiKeyMap DriverId ModuleName ->
+  TimingUI ->
+  TimingTable ->
+  Widget IHTML Event
+renderMemChart drvModMap tui ttable =
+  S.svg
+    svgProps
+    [ S.style [] [text ".small { font: 5px sans-serif; } text { user-select: none; }"]
+    , S.g [] (fmap (renderPrimitive (const [])) rexp)
+    ]
+  where
+    scene = TimingView.compileMemChart drvModMap tui ttable
+    rexp = sceneElements scene
+    vpi = tui ^. timingUIViewPort
+    vp = fromMaybe (vpi ^. vpViewPort) (vpi ^. vpTempViewPort)
+    viewboxProp =
+      SP.viewBox . T.intercalate " " . fmap (T.pack . show . floor @_ @Int) $
+        [ 0
+        , topLeft vp ^. _2
+        , 300
+        , timingHeight
+        ]
+    svgProps =
+      [ width "300"
+      , height (T.pack (show (timingHeight :: Int)))
+      , viewboxProp
+      , SP.version "1.1"
+      , xmlns
+      ]
+
+renderTimingRange ::
+  TimingUI ->
+  TimingTable ->
+  Widget IHTML Event
+renderTimingRange tui ttable =
+  div [] [svgElement]
+  where
+    scene = TimingView.compileTimingRange tui ttable
+    rexp = sceneElements scene
+    svgProps =
+      [ width (T.pack (show (timingWidth :: Int)))
+      , height (T.pack (show (timingRangeHeight :: Int)))
+      , SP.version "1.1"
+      , xmlns
+      ]
+    svgElement =
+      S.svg
+        svgProps
+        [ S.style [] [text ".small { font: 5px sans-serif; } text { user-select: none; }"]
+        , S.g [] (fmap (renderPrimitive (const [])) rexp)
+        ]
+
+renderBlockers :: ModuleName -> TimingTable -> Widget IHTML Event
+renderBlockers hoveredMod ttable =
+  divClass "blocker" [] [selected, upstream, hr [], downstreams]
+  where
+    upMods =
+      maybeToList (M.lookup hoveredMod (ttable ^. ttableBlockingUpstreamDependency))
+    downMods =
+      fromMaybe [] (M.lookup hoveredMod (ttable ^. ttableBlockedDownstreamDependency))
+
+    selected =
+      divClass "box" [] [p [] [text hoveredMod]]
+    upstream =
+      div
+        []
+        ( divClass "blocker title" [] [text "blocked by"]
+            : fmap (\modu -> p [] [text modu]) upMods
+        )
+    downstreams =
+      div
+        []
+        ( divClass "blocker title" [] [text "blocking"]
+            : fmap (\modu -> p [] [text modu]) downMods
+        )
+
+renderTimingView ::
+  BiKeyMap DriverId ModuleName ->
+  TimingUI ->
+  TimingTable ->
+  Widget IHTML Event
+renderTimingView drvModMap tui ttable =
+  divClass
+    "box"
+    []
+    ( [ divClass
+          "columns"
+          []
+          [ divClass
+              "box column is-four-fifths"
+              [style [("overflow", "hidden")]]
+              [ renderTimingChart drvModMap tui ttable
+              ]
+          , divClass
+              "box column is-one-fifth"
+              [style [("overflow", "hidden")]]
+              [renderMemChart drvModMap tui ttable]
+          ]
+      , renderTimingRange tui ttable
+      ]
+        ++ hoverInfo
+    )
+  where
+    mhoveredMod = tui ^. timingUIHoveredModule
+    hoverInfo =
+      case mhoveredMod of
+        Nothing -> []
+        Just hoveredMod ->
+          [ divClass
+              "box"
+              [ style
+                  [ ("width", "150px")
+                  , ("height", "120px")
+                  , ("position", "absolute")
+                  , ("bottom", "0")
+                  , ("left", "0")
+                  , ("background", "ivory")
+                  , ("overflow", "hidden")
+                  ]
+              ]
+              [renderBlockers hoveredMod ttable]
+          ]
 
 buttonShowBlocker :: TimingUI -> Widget IHTML Event
 buttonShowBlocker tui = divClass "control" [] [button']
@@ -146,7 +342,7 @@ renderTimingMode model ss =
             , ("position", "relative")
             ]
         ]
-        ( [ TimingView.render (ss ^. serverDriverModuleMap) (model ^. modelTiming) ttable
+        ( [ renderTimingView (ss ^. serverDriverModuleMap) (model ^. modelTiming) ttable
           , div
               [ style
                   [ ("position", "absolute")
